Homework_14_1/requirements.txt
aiosmtplib==2.0.2
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.14.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-limiter==0.1.6
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
patchlib==0.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==1.3.13
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.0.post1
